Check in SkunkBoard User Manual v1.0 files
This includes all the files used to construct
the manual, including the exact PDF submitted to
the printer (manual/releases/skunk_manual-1.0.pdf)

Also included are two scripts that can be used to
build the manual PDF from the command line on
Linux in both a printshop-friendly basic format,
as well as a print-at-home booklet format. These
scripts require various support programs. On
Ubuntu, this should be sufficient to install all
of them:

 # apt install texlive-extra-utils libreoffice